Move error types to internal/errors package
